FirstTask/requirements.txt
asyncssh==2.19.0
cffi==1.17.1
cryptography==44.0.0
greenlet==3.1.1
playwright==1.50.0
pycparser==2.22
pyee==12.1.1
typing_extensions==4.12.2
paho-mqtt>=1.6.1